+++ tcod/ecs/__init__.py
@@ -45,7 +45,16 @@ _ComponentKey = Union[Type[T], Tuple[object, Type[T]]]
 
 
 def abstract_component(cls: type[T]) -> type[T]:
-    """Register class `cls` as an abstract component and return it."""
+    """Register class `cls` as an abstract component and return it.
+
+    .. deprecated:: Unreleased
+        This decorator is deprecated since abstract components should always be explicit.
+    """
+    warnings.warn(
+        "This decorator is deprecated since abstract components should always be explicit.",
+        FutureWarning,
+        stacklevel=2,
+    )
     cls._TCOD_BASE_COMPONENT = cls  # type: ignore[attr-defined]
     return cls
 
@@ -65,9 +74,11 @@ class Entity:
     Example::
 
         >>> import tcod.ecs
-        >>> world = tcod.ecs.World()  # Create a new world.
-        >>> entity = world.new_entity(name="entity")  # Create a new entity, name is optional.
-        >>> other_entity = world.new_entity(name="other_entity")
+        >>> world = tcod.ecs.World()  # Create a new world
+        >>> world.new_entity()  # Create a new entity
+        <Entity(uid=object at ...)>
+        >>> entity = world["entity"]  # Get an entity from a specific identifier
+        >>> other_entity = world["other"]
     """  # Changes here should be reflected in conftest.py.
 
     __slots__ = ("world", "uid", "__weakref__")
@@ -125,9 +136,9 @@ class Entity:
             >>> {str, ("name", str)}.issubset(entity.components.keys())
             True
             >>> list(world.Q.all_of(components=[str]))  # Query components.
-            [<Entity name='entity'>]
+            [<Entity(uid='entity')>]
             >>> list(world.Q[tcod.ecs.Entity, str, ("name", str)])  # Query zip components.
-            [(<Entity name='entity'>, 'foo', 'my_name')]
+            [(<Entity(uid='entity')>, 'foo', 'my_name')]
         """
         return EntityComponents(self)
 
@@ -141,7 +152,7 @@ class Entity:
             >>> "tag" in entity.tags  # Check tag.
             True
             >>> list(world.Q.all_of(tags=["tag"]))  # Query tags.
-            [<Entity name='entity'>]
+            [<Entity(uid='entity')>]
             >>> entity.tags.discard("tag")
         """
         return EntityTags(self)
@@ -157,13 +168,13 @@ class Entity:
             >>> other_entity in entity.relation_components[str]
             True
             >>> list(world.Q.all_of(relations=[(str, other_entity)]))
-            [<Entity name='entity'>]
+            [<Entity(uid='entity')>]
             >>> list(world.Q.all_of(relations=[(str, ...)]))
-            [<Entity name='entity'>]
+            [<Entity(uid='entity')>]
             >>> list(world.Q.all_of(relations=[(entity, str, None)]))
-            [<Entity name='other_entity'>]
+            [<Entity(uid='other')>]
             >>> list(world.Q.all_of(relations=[(..., str, None)]))
-            [<Entity name='other_entity'>]
+            [<Entity(uid='other')>]
         """
         return EntityComponentRelations(self)
 
@@ -175,9 +186,9 @@ class Entity:
 
             >>> entity.relation_tags["ChildOf"] = other_entity  # Assign relation.
             >>> list(world.Q.all_of(relations=[("ChildOf", other_entity)]))  # Get children of other_entity.
-            [<Entity name='entity'>]
+            [<Entity(uid='entity')>]
             >>> list(world.Q.all_of(relations=[(entity, "ChildOf", None)]))  # Get parents of entity.
-            [<Entity name='other_entity'>]
+            [<Entity(uid='other')>]
             >>> del entity.relation_tags["ChildOf"]
         """
         return EntityRelationsExclusive(self)
@@ -192,16 +203,12 @@ class Entity:
         """
         return EntityRelations(self)
 
-    @property
-    def name(self) -> object:
-        """The unique name of this entity or None.
-
-        You may assign a new name, but if an entity of the world already has that name then it will lose it.
-        """
-        return self.world._names_by_entity.get(self)
-
-    @name.setter
-    def name(self, value: object) -> None:
+    def _set_name(self, value: object, stacklevel: int = 1) -> None:
+        warnings.warn(
+            "The name feature has been deprecated and will be removed.",
+            FutureWarning,
+            stacklevel=stacklevel + 1,
+        )
         old_name = self.name
         if old_name is not None:  # Remove self from names.
             del self.world._names_by_name[old_name]
@@ -214,6 +221,21 @@ class Entity:
             self.world._names_by_name[value] = self
             self.world._names_by_entity[self] = value
 
+    @property
+    def name(self) -> object:
+        """The unique name of this entity or None.
+
+        You may assign a new name, but if an entity of the world already has that name then it will lose it.
+
+        .. deprecated:: Unreleased
+            This feature has been deprecated.
+        """
+        return self.world._names_by_entity.get(self)
+
+    @name.setter
+    def name(self, value: object) -> None:
+        self._set_name(value, stacklevel=2)
+
     def __repr__(self) -> str:
         """Return a representation of this entity.
 
@@ -223,8 +245,6 @@ class Entity:
             <Entity(uid=object at ...)>
             >>> world["foo"]
             <Entity(uid='foo')>
-            >>> world.new_entity(name="foo")
-            <Entity name='foo'>
         """
         uid_str = f"object at 0x{id(self.uid):X}" if self.uid.__class__ == object else repr(self.uid)
         items = [f"{self.__class__.__name__}(uid={uid_str})"]
@@ -258,8 +278,17 @@ class EntityComponents(MutableMapping[Union[Type[Any], Tuple[object, Type[Any]]]
         """Initialize this attribute for the given entity."""
         self.entity: Final = entity
 
-    def set(self, value: object) -> None:
-        """Assign or overwrite a component, automatically deriving the key."""
+    def set(self, value: object, *, _stacklevel: int = 1) -> None:
+        """Assign or overwrite a component, automatically deriving the key.
+
+        .. deprecated:: Unreleased
+            Setting values without an explicit key has been deprecated.
+        """
+        warnings.warn(
+            "Setting values without an explicit key has been deprecated.",
+            FutureWarning,
+            stacklevel=_stacklevel + 1,
+        )
         key = getattr(value, "_TCOD_BASE_COMPONENT", value.__class__)
         self[key] = value
 
@@ -308,23 +337,24 @@ class EntityComponents(MutableMapping[Union[Type[Any], Tuple[object, Type[Any]]]
         """Return the number of components belonging to this entity."""
         return len(self.entity.world._components_by_entity.get(self.entity, ()))
 
-    def update_values(self, values: Iterable[object]) -> None:
-        """Add or overwrite multiple components inplace, deriving the keys from the values."""
+    def update_values(self, values: Iterable[object], *, _stacklevel: int = 1) -> None:
+        """Add or overwrite multiple components inplace, deriving the keys from the values.
+
+        .. deprecated:: Unreleased
+            Setting values without an explicit key has been deprecated.
+        """
         for value in values:
-            self.set(value)
+            self.set(value, _stacklevel=_stacklevel + 1)
 
     def by_name_type(self, name_type: type[_T1], component_type: type[_T2]) -> Iterator[tuple[_T1, type[_T2]]]:
         """Iterate over all of an entities component keys with a specific (name_type, component_type) combination.
 
-        .. versionadded:: 3.0.0
+        .. versionadded:: 3.0
 
-        Example::
-
-            >>> entity.components["A", int] = 1
-            >>> entity.components["B", int] = 2
-            >>> sorted(entity.components.by_name_type(str, int))
-            [('A', <class 'int'>), ('B', <class 'int'>)]
+        .. deprecated:: Unreleased
+            This method has been deprecated. Iterate over items instead.
         """
+        warnings.warn("This method has been deprecated. Iterate over items instead.", FutureWarning, stacklevel=2)
         # Naive implementation until I feel like optimizing it.
         for key in self:
             if not isinstance(key, tuple):
@@ -763,34 +793,47 @@ class World:
     def named(self) -> Mapping[object, Entity]:
         """A view into this worlds named entities.
 
-        Example::
-
-            >>> entity = world.new_entity(name="MyEntity")  # Name is optional and can be any hashable, not just str.
-            >>> entity
-            <Entity name='MyEntity'>
-            >>> entity.world is world  # Worlds can always be accessed from their entity.
-            True
-            >>> world.named["MyEntity"]  # Named entities can be accessed with `World.named`.
-            <Entity name='MyEntity'>
-            >>> world.named.get("Missing") is None  # `World.named` acts like a dictionary, including assignment and `.get()`.
-            True
+        .. deprecated:: Unreleased
+            This feature has been deprecated.
         """
         return self._names_by_name
 
     def new_entity(
         self,
-        components: Iterable[object] = (),
+        components: Iterable[object] | Mapping[_ComponentKey[object], object] = (),
         *,
         name: object = None,
         tags: Iterable[Any] = (),
     ) -> Entity:
-        """Create and return a new entity."""
+        """Create and return a new entity.
+
+        .. versionchanged:: Unreleased
+            `components` can now take a mapping.
+
+        Example::
+
+            >>> entity = world.new_entity(
+            ...     components={
+            ...         ("name", str): "my name",
+            ...         ("hp", int): 10,
+            ...     },
+            ...     tags=["Actor"],
+            ... )
+            >>> entity.components[("name", str)]
+            'my name'
+            >>> "Actor" in entity.tags
+            True
+        """
         entity = Entity(self)
-        entity.components.update_values(components)
+        if isinstance(components, Mapping):
+            entity.components.update(components)
+        elif components:
+            entity.components.update_values(components, _stacklevel=2)
         entity_tags = entity.tags
         for tag in tags:
             entity_tags.add(tag)
-        entity.name = name
+        if name is not None:
+            entity._set_name(name, stacklevel=2)
         return entity
 
     @property